Move auth success to the homepage
