+++ apps/client/src/main.tsx
@@ -6,7 +6,6 @@ import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import AuthCallbackPage from "./routes/auth/AuthCallbackPage";
 import LoginPage from "./routes/auth/LoginPage";
 import RegisterPage from "./routes/auth/RegisterPage.tsx";
-import AuthSuccessPage from "./routes/auth/AuthSuccessPage.tsx";
 import { AuthContextProvider } from "./context/authContext.tsx";
 
 const router = createBrowserRouter([
@@ -14,7 +13,6 @@ const router = createBrowserRouter([
   { path: "/auth/login", element: <LoginPage /> },
   { path: "/auth/register", element: <RegisterPage /> },
   { path: "/auth/callback", element: <AuthCallbackPage /> },
-  { path: "/auth/success", element: <AuthSuccessPage /> },
   { path: "*", element: <div>404</div> },
 ]);
